refactor: use custom hook instead of effect in component
