+++ packages/esm-billing-app/translations/en.json
@@ -77,7 +77,6 @@
   "loadingDescription": "Loading",
   "makeclaims": "Make Claims",
   "manageBillableServices": "Manage billable services",
-  "mpesaError": "Mpesa Error",
   "name": "Name",
   "navigateBack": "Navigate back",
   "nextPage": "Next page",
@@ -133,8 +132,6 @@
   "shaNumber": "SHA Number",
   "shortName": "Short Name",
   "status": "Service Status",
-  "stkPush": "STK Push",
-  "stkPushSucess": "STK Push send successfully",
   "stockItem": "Stock Item",
   "total": "Total",
   "totalAmount": "Total Amount",
